generalise terratorch models support via a dedicated backend
